wealthra-stt/requirements.txt
groq>=0.5.0
python-dotenv>=1.0.1
fastapi
uvicorn
python-multipart